requirements.txt
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
cookiecutter==1.7.3
coverage==6.3
cssselect==1.1.0
Django==3.2.10
django-bootstrap-pagination==1.7.1
django-browser-reload==1.3.0
django-countries==7.2.1
django-scheduler==0.9.5
django-tailwind==3.1.1
django-widget-tweaks==1.4.12
django-with-asserts==0.0.1
Faker==11.3.0
humanize==3.14.0
icalendar==4.0.9
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
joblib==1.1.0
libgravatar==1.0.0
lxml==4.7.1
MarkupSafe==2.0.1
numpy==1.22.1
pandas==1.4.0
Pillow==9.0.1
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
pyuca==1.2
requests==2.27.1
scikit-learn==1.0.2

scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
urllib3==1.26.8
